+++ Frameworks/PortMapper/framework/libnatpmp/natpmp.h
@@ -177,7 +177,7 @@ LIBSPEC int getnatpmprequesttimeout(natpmp_t * p, struct timeval * timeout);
  * NATPMP_ERR_OUTOFRESOURCES
  * NATPMP_ERR_UNSUPPORTEDOPCODE
  * NATPMP_ERR_UNDEFINEDERROR */
-LIBSPEC int readnatpmpresponseorretry(natpmp_t * p, natpmpresp_t * response);
+LIBSPEC ssize_t readnatpmpresponseorretry(natpmp_t * p, natpmpresp_t * response);
 
 #ifdef ENABLE_STRNATPMPERR
 LIBSPEC const char * strnatpmperr(int t);
